requirements.txt
streamlit==1.26.0
PyPDF2==3.0.1
langchain==0.1.1
langchain-community==0.0.13
chromadb==0.4.22
pydantic==2.5.3
nltk==3.8.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.33.2
faiss-cpu==1.8.0.post1
langchain-groq==0.0.1
pdf2image==1.17.0
pytesseract==0.3.13